Refactor ENV loading into dedicated package
